requirements.txt
torch
transformers
datasets
openai>=1.0.0
python-dotenv
tqdm
numpy 